Adds indexing to (object type, object name), instead of just object name
This allows for addressing overloading names of differing types and proper
indexing to support all of this. This cleans up some of the indexing and updates
tests back to calling the long object type name, which is easier to use from
tests as a reference to the short reference type name.

This fixes #15