Fix any reference lookups on domain
This limits any reference lookups to only the main roles, designating roles with long directives
as alternate roles. The default implementation here is for the Sphinx env to lookup against all
roles on the domain, which will return duplicate objects, as both the short and long directives
point to the same object class. Ensure we have no warnings when using using any references that
shouldn't collide.